requirements.txt
django
pillow
openpyxl
django-impersonate
django-maintenancemode
django-render-block
django-redis
channels
asgi_redis
daphne
psycopg2
docutils
django-ipware
django-sendfile
django-two-factor-auth
xhtml2pdf>=0.2b1
django-htmlmin

Sphinx

djangosaml2
service_identity